requirements.txt
crewai>=0.1.0
google-generativeai==0.3.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
streamlit==1.28.1
pandas==2.1.3
python-dotenv==1.0.0
youtube-transcript-api==0.6.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
